feat: add verifyUser + booking tests

+++ src/controllers/booking.controller.js
@@ -3,6 +3,7 @@ const pick = require('../utils/pick');
 const ApiError = require('../utils/ApiError');
 const catchAsync = require('../utils/catchAsync');
 const { bookingService } = require('../services');
+const verifyUser = require('../utils/verifyUser');
 
 const getBookings = catchAsync(async (req, res) => {
   const filter = pick(req.query, ['parkingLotId', 'clientId', 'isCancelled']);
@@ -16,10 +17,16 @@ const getBookingById = catchAsync(async (req, res) => {
   if (!booking) {
     throw new ApiError(httpStatus.NOT_FOUND, 'Booking not found');
   }
+  await verifyUser(req.user, booking.clientId);
   res.send(booking);
 });
 
 const deleteBookingById = catchAsync(async (req, res) => {
+  const booking = await bookingService.getBookingById(req.params.bookingId);
+  if (!booking) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Booking not found');
+  }
+  await verifyUser(req.user, booking.clientId);
   await bookingService.deleteBookingById(req.params.bookingId);
   res.status(httpStatus.NO_CONTENT).send();
 });
